FreeBSD: Added DHCP start/stop functionality to FreeBSD Distro
Added stopDHCP and startDHCP FreeBSD functions.  Also added a dynamic
isDHCPEnabled FreeBSD function which queries if the dhclient service is
running.  Modified checks on DoDhcpWork to restart the DHCP client if it was
previously running when the Distro object used a isDHCPEnabled function that
checks if the DHCP client is running.